+++ HW1/Search/Heuristic.java
@@ -3,5 +3,5 @@ package Search;
 import Problem.State;
 
 public interface Heuristic {
-    int calculate(State state);
+    long calculate(State state);
 }
